feat: Add implementation of interweave
This commit adds a first draft of the implementation. It is considered
to be mainly complete. The algorithm is O(N), the type checking is in
place and the first tests pass.

What is left is to do some refactoring to improve the code base and to
extend the test coverage.