feat: Handle boundary conditions better
There may be events E1 and E2 where E2 begins exactly when E1 starts.
This case was not covered by tests before. The agreed on behaviour is to
assume that E2 begins a tiny bit after E1, so that the do not appear
simultaneously.

This commit adds a test and a bit of documentation for that behaviour.